[compat, sample, ->RELENG_1_2] refs #4398
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist-Python/trunk@907 3ceb7c45-664a-4a73-a5f3-5d010aa2b7c3

+++ OpenRTM_aist/examples/Composite/Motor.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 # -*- Python -*-
 
+from __future__ import print_function
 import sys
 import time
 sys.path.append(".")
@@ -51,8 +52,8 @@ class Motor(OpenRTM_aist.DataFlowComponentBase):
   def onExecute(self, ec_id):
     if self._inIn.isNew():
       data = self._inIn.read()
-      print "Motor Received data: ", data.data
-      self._d_out.data = long(data.data *2)
+      print("Motor Received data: ", data.data)
+      self._d_out.data = int(data.data *2)
       self._outOut.write()
     return RTC.RTC_OK
